examples/python/requirements.txt
msgpack>=1.0.0
pyzmq>=23.0.0
pydantic>=1.10.4
typing_extensions>=4.4.0
